refactor: 파일명, addCoin, 에러핸들링 수정
